feat: update to validate params in new API
