--- src/frame/auction/types.js
@@ -1,10 +1,0 @@
-export default {
-	seller: 'string',
-	decision_logic_url: 'url',
-	interest_group_buyers: 'mixed',
-	trusted_scoring_signals_url: 'url',
-	additional_bids: 'array',
-	auction_signals: 'object',
-	seller_signals: 'object',
-	per_buyer_signals: 'object',
-};
