+++ src/frame/interest-groups/index.js
@@ -1,16 +1,8 @@
-import { db, echo, validate } from '../../utils/index.js';
+import { db, echo } from '../../utils/index.js';
 import { IG_STORE } from '../../utils/db.js';
-import types from './types.js';
 import {
 	getIGKey,
 } from './utils.js';
-
-/*
- * @const {number}
- * @summary Maximum expiration allowed for an Interest Group to exist
- * @description Milliseconds occuring per day multiplied by the maximum number of days (maximum dayus (30) * hours per day (24) * minutes per hour (60) * seconds per minute (60) * milliseconds per second (1000))
- */
-const MAX_EXPIRATION = 2592000000;
 
 /*
  * @function
@@ -25,19 +17,7 @@ const MAX_EXPIRATION = 2592000000;
  * @example
  *   joinAdInterestGroup({ owner: 'foo', name: 'bar', bidding_logic_url: 'http://example.com/bid' }, 2592000000);
  */
-export default async function joinAdInterestGroup (options, expiry, debug = false) {
-	debug && echo.group('Fledge: Join an Interest Group');
-	debug && echo.log('interest group options:', options);
-	debug && echo.log('interest group expiration:', expiry);
-	validate.param(options, 'object');
-	validate.param(expiry, 'number');
-	validate.hasRequiredKeys(options, [ 'owner', 'name', 'bidding_logic_url' ]);
-	validate.hasInvalidOptionTypes(options, types);
-
-	if (expiry > MAX_EXPIRATION) {
-		throw new Error(`'expiry' is set past the allowed maximum value. You must provide an expiration that is less than or equal to ${MAX_EXPIRATION}.`);
-	}
-
+export async function joinAdInterestGroup (options, expiry, debug) {
 	debug && echo.info('checking for an existing interest group');
 	const group = await db.store.get(IG_STORE, getIGKey(options.owner, options.name));
 	debug && echo.table(group);
@@ -57,6 +37,27 @@ export default async function joinAdInterestGroup (options, expiry, debug = fals
 		});
 	}
 	debug && echo.log('interest group id:', id);
+	debug && echo.groupEnd();
+
+	return true;
+}
+
+/*
+ * @function
+ * @name leaveAdInterestGroup
+ * @description leave an interest group removing from IndexedDB
+ * @author Newton <cnewton@magnite.com>
+ * @param {object} options - An object of options to create an interest group {@link types}
+ * @throws {Error} Any parameters passed are incorrect or an incorrect type
+ * @return {true}
+ *
+ * @example
+ *   leaveAdInterestGroup({ owner: 'foo', name: 'bar', bidding_logic_url: 'http://example.com/bid' });
+ */
+export async function leaveAdInterestGroup (group, debug) {
+	debug && echo.info('deleting an existing interest group');
+	await db.store.delete(IG_STORE, getIGKey(group.owner, group.name));
+	debug && echo.log('interest group deleted');
 	debug && echo.groupEnd();
 
 	return true;
